+++ backend/srcs/src/game/game.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { GameService } from './game.service';
 import { GameGateway } from './game.gateway';
 import { HistoricGamesService } from 'src/historic_games/historic_games.service';
+import { StatsService } from 'src/stats/stats.service';
 import { AuthService } from 'src/auth/auth.service';
 import { JwtService } from "@nestjs/jwt";
 import { UserService } from 'src/user/user.service';
@@ -11,6 +12,7 @@ import { UserService } from 'src/user/user.service';
     GameService,
     GameGateway,
     HistoricGamesService,
+    StatsService,
     AuthService,
     JwtService, 
     UserService