+++ frontend/srcs/src/app/services/index.ts
@@ -4,3 +4,4 @@ export * from './chat.service';
 export * from './channel.service';
 export * from './user.service';
 export * from './game.service';
+export * from './pongtv.service'